Refactor database into a separate file
