+++ main.py
@@ -8,10 +8,10 @@ Main! yay!
 import logging
 
 
-from peewee import BigIntegerField, CharField, IntegerField, Model, SqliteDatabase
 from telegram.ext import Updater
 
 
+from databases import User
 from modules.bot import Bot
 from modules.community.exp import Exp
 from modules.community.karma import Karma
@@ -29,53 +29,6 @@ logging.basicConfig(
 )
 
 logger = logging.getLogger(__name__)
-
-# Database
-main_db = SqliteDatabase("./databases/main.db")
-
-
-class User(Model):
-    """
-    User model to access the database.
-    """
-
-    userid = BigIntegerField()
-    userfirstname = CharField(null=True)
-    chatid = BigIntegerField()
-    karma = IntegerField(default=0)
-    num_messages = IntegerField(default=0)
-    level = IntegerField(default=0)
-
-    class Meta:
-        """
-        Basically which database.
-        """
-
-        database = main_db
-
-
-class ChatModule(Model):
-    """
-    Chat command model to access the database.
-    """
-
-    chatid = BigIntegerField()
-    commandname = CharField()
-    enabled = IntegerField(default=1)
-
-
-class ChatCommand(Model):
-    """
-    Chat command model to access the database.
-    """
-
-    chatid = BigIntegerField()
-    modulename = CharField()
-    enabled = IntegerField(default=1)
-
-
-main_db.connect()
-main_db.create_tables([User, ChatModule, ChatCommand])
 
 
 def main() -> None:
