+++ requirements.txt
@@ -1,5 +1,5 @@
 beautifulsoup4==4.10.0
-black==21.6b0
+#black==21.6b0
 dateparser==1.0.0
 isort==5.8.0
 peewee==3.14.4
